+++ src/debugger/datatip.jl
@@ -1,0 +1,31 @@
+using CodeTracking
+using JuliaInterpreter: getfile, locals
+import ..Atom: strlimit
+
+function datatip(word, path, row, column, s::DebuggerState = STATE)
+  frame = s.frame
+
+  # frame validation & only work for methods
+  scope = frame.framecode.scope
+  if frame === nothing || scope isa Module
+    return nothing
+  end
+
+  # path identity check
+  path != getfile(frame) && return nothing
+
+  # line number check
+  defstr, startline = definition(String, scope)
+  endline = startline + sum(c === '\n' for c in defstr)
+  startline <= row <= endline || return nothing
+
+  # local bindings
+  for v in locals(frame)
+    if string(v.name) === word
+      valstr = @> repr(MIME("text/plain"), v.value, context = :limit => true) strlimit(1000)
+      return [Dict(:type  => :snippet, :value => valstr)]
+    end
+  end
+
+  return nothing # when no local binding exists
+end
